Move user service logic to model
