+++ src/controller/clientController.js
@@ -1,5 +1,5 @@
 import * as clientService from '../service/clientService.js';
-import * as userService from '../service/userService.js';
+import * as userModel from '../model/userModel.js';
 import * as instaPostService from '../service/instaPostService.js';
 import * as instaLikeService from '../service/instaLikeService.js';
 import * as tiktokPostService from '../service/tiktokPostService.js';
@@ -68,7 +68,7 @@ export const deleteClient = async (req, res, next) => {
 // Semua user di bawah client
 export const getUsers = async (req, res, next) => {
   try {
-    const users = await userService.findUsersByClientId(req.params.client_id);
+    const users = await userModel.findUsersByClientId(req.params.client_id);
     sendSuccess(res, users);
   } catch (err) {
     next(err);
@@ -140,7 +140,7 @@ export const getSummary = async (req, res, next) => {
     const client = await clientService.findClientById(client_id);
     if (!client) return res.status(404).json({ error: 'Client not found' });
 
-    const users = await userService.findUsersByClientId(client_id);
+    const users = await userModel.findUsersByClientId(client_id);
     const instaPosts = await instaPostService.findByClientId(client_id);
     let instaLikes = 0;
     for (const post of instaPosts) {
